Changed Locations and Method Names
