--- Data Access Layer/FileHandler.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Group_Project_PRG282.Data_Access_Layer
-{
-    class FileHandler
-    {
-    }
-}
