+++ PresentationLayer/frmLogin.resx
