+++ PresentationLayer/frmMain.resx
