+++ PresentationLayer/frmRegister.resx
